Enemy and Collision

+++ CraftEngine/Util/Util.h
@@ -23,7 +23,7 @@ namespace Util
 	}
 
 	// 정수 난수 함수
-	inline float RandomRange(int min, int max)
+	inline int RandomRange(int min, int max)
 	{
 		// min에서 max까지 균등하게 부동소수점 난수를 생성하주는 분포 정의
 		std::uniform_int_distribution<int> distribution(min, max);
